Project App/requirements.txt
asgiref==3.5.2
aliyun-python-sdk-core==2.13.3
async-timeout==4.0.2
Deprecated==1.2.13
Django==4.0.6
django-crispy-forms==1.14.0
django-defender==0.9.5
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
djangorestframework==3.13.1
django-environ==0.9.0
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
gunicorn==20.1.0
pytz==2022.1
redis==4.3.4
sqlparse==0.4.2
tzdata==2022.1
twilio==7.15.4
wrapt==1.14.1
django-filter==22.1
django-on-heroku==1.1.2
tensorflow==2.12.0
transformers==4.26.1
numpy==1.22.4
matplotlib==3.6.3
matplotlib-inline==0.1.6